LocalVLLMModel and deployment handler (#102)
This PR introduces the `LocalVLLMModel` and
`LocalVLLMDeploymentHandler`, the former is a usual Model (i.e.
implements generate) and the latter is a new class to handle aspects of
deployment (deployment itself, health checks on your deployment,
shutdown of servers).

Use LocalVLLMModel either by defining a ModelConfig or by passing info
in the command line so that vllm can recognize your deployment or even
deploy for you.

If you have already deployed, pass "ports" parameter, otherwise the
LocalVLLMDeploymentHandler will spin up "num_servers" (default = 1)
servers for you, wait for deployment to finish, and continue with the
eval pipeline.

---------

Co-authored-by: Michael Harrison <mharrison@microsoft.com>
Co-authored-by: Vidhisha Balachandran <vidhishab@microsoft.com>