Remove inference batching (#111)
Removed batching, swapped some logic of the dataloader /
ThreadPoolExecutor / writer in inference.py, refactored the _run
methods.

Current flow is: when run is called, load partial results if
appropriate, then kick off ThreadPoolExecutor, which calls "_run_single"
for each element of the dataloader and then appends the results to the
file (note that appending vs writing is new -- let me know if this is
not preferred for any reason). _run_single is a combination of the
previous _run and _run_par: it checks for previous results, checks for
rate limiting, and finally calls the model's generate().

---------

Co-authored-by: Safoora Yousefi <sayouse@microsoft.com>
Co-authored-by: Michael Harrison <mharrison@microsoft.com>