minor fixes (#117)
Three small fixes:
1) update gitignore to ignore files from editable install

2) fixing issue where logs contain api_key. since the __repr__ from
dataclass was overriding the custom one from UtilityClassConfig, the
secret_key_params and api_key were not getting scrubbed before logging.

3) isna() doesn't always return a bool (e.g. if it's called on something
arraylike, it returns a series or dataframe of bools, which causes a
ValueError here). the fix i've proposed will replace NaN things with
self.value, but won't do anything with arraylike things for which some
element is NaN (since in that case, it's unclear whether the user wants
to replace the whole array with self.value or replace the offending
elements with NaN, and I think in this case we should let it throw so
that they can decide for themselves. in any case i don't really expect
eval datasets to contain arrays like [5, NaN, "hello"], so this is
probably a non-issue).

Co-authored-by: Michael Harrison <mharrison@microsoft.com>